Fixed infinite fetch in CourseDetail component
I decided that the best way to fix the infinite fetch was to not use the getCourse method built in the Data class as I needed to make sure the location match the course view requested.  It was easier to do this by just fetching the data directly in the CourseDetail component.